Move metrics reporting into its own event handler.
This breaks the "state keeping" of the KnativeServing reconciler out of the reconciler and pushes it into a stateless event-handler format. Also moves the stats reporter code into a KnativeServing specific package.
