+++ pkg/reconciler/knativeserving/controller.go
@@ -15,6 +15,7 @@ package knativeserving
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"path/filepath"
 
@@ -30,10 +31,12 @@ import (
 	mf "github.com/manifestival/manifestival"
 	"k8s.io/client-go/tools/cache"
 	"knative.dev/operator/pkg/apis/operator/v1alpha1"
+	servingv1alpha1 "knative.dev/operator/pkg/apis/operator/v1alpha1"
 	knativeServinginformer "knative.dev/operator/pkg/client/injection/informers/operator/v1alpha1/knativeserving"
 	knsreconciler "knative.dev/operator/pkg/client/injection/reconciler/operator/v1alpha1/knativeserving"
 	"knative.dev/operator/pkg/reconciler"
 	"knative.dev/operator/pkg/reconciler/common"
+	servingcommon "knative.dev/operator/pkg/reconciler/knativeserving/common"
 	deploymentinformer "knative.dev/pkg/client/injection/kube/informers/apps/v1/deployment"
 	"knative.dev/pkg/configmap"
 	"knative.dev/pkg/controller"
@@ -56,7 +59,7 @@ func NewController(ctx context.Context, cmw configmap.Watcher) *controller.Impl 
 		logger.Fatalw("Failed to remove old resources", zap.Error(err))
 	}
 
-	statsReporter, err := reconciler.NewStatsReporter(controllerAgentName)
+	statsReporter, err := servingcommon.NewStatsReporter(controllerAgentName)
 	if err != nil {
 		logger.Fatal(err)
 	}
@@ -64,7 +67,6 @@ func NewController(ctx context.Context, cmw configmap.Watcher) *controller.Impl 
 	c := &Reconciler{
 		kubeClientSet:           kubeClient,
 		knativeServingClientSet: servingclient.Get(ctx),
-		statsReporter:           statsReporter,
 		knativeServingLister:    knativeServingInformer.Lister(),
 		servings:                map[string]int64{},
 		platform:                common.GetPlatforms(ctx),
@@ -91,5 +93,30 @@ func NewController(ctx context.Context, cmw configmap.Watcher) *controller.Impl 
 		Handler:    controller.HandleAll(impl.EnqueueControllerOf),
 	})
 
+	// Reporting statistics on KnativeServing events.
+	knativeServingInformer.Informer().AddEventHandler(cache.ResourceEventHandlerFuncs{
+		AddFunc: func(newObj interface{}) {
+			new := newObj.(*servingv1alpha1.KnativeServing)
+			if new.Generation == 1 {
+				statsReporter.ReportKnativeservingChange(key(new), "creation")
+			}
+		},
+		UpdateFunc: func(oldObj interface{}, newObj interface{}) {
+			old := oldObj.(*servingv1alpha1.KnativeServing)
+			new := newObj.(*servingv1alpha1.KnativeServing)
+			if old.Generation < new.Generation {
+				statsReporter.ReportKnativeservingChange(key(new), "edit")
+			}
+		},
+		DeleteFunc: func(oldObj interface{}) {
+			old := oldObj.(*servingv1alpha1.KnativeServing)
+			statsReporter.ReportKnativeservingChange(key(old), "deletion")
+		},
+	})
+
 	return impl
+}
+
+func key(ks *servingv1alpha1.KnativeServing) string {
+	return fmt.Sprintf("%s/%s", ks.Namespace, ks.Name)
 }
